Move branding place and create config
