+++ app/app.html
@@ -8,7 +8,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" />
 	<link rel="stylesheet" type="text/css" href="./stylesheets/base.css">
 	<link rel="stylesheet" type="text/css" href="./stylesheets/start-page.css">
-	<link rel="stylesheet" type="text/css" href="./branding/css/branding.css">
+	<link rel="stylesheet" type="text/css" href="./branding/branding.css">
 </head>
 <body class="hide-server-list">
 	<div class="server-list">
