+++ app/app.js
@@ -1,4 +1,4 @@
-import './branding/js/branding.js';
+import './branding/branding.js';
 import { start } from './scripts/start';
 
 window.$ = window.jQuery = require('./vendor/jquery-1.12.0');
